requirements.txt
python-telegram-bot==10.0.1
requests==2.18.4
